refactor log

+++ src/presenter/sensor_presenter.py
@@ -1,5 +1,4 @@
 from src.model.sensor import FlexSensorData, ForceSensorData
-from src.util.sensor_log import SensorLog
 
 class SensorPresenter:
     """Presenter for handling sensor data"""
@@ -18,7 +17,7 @@ class SensorPresenter:
         self.view.loop = loop  # Set event loop for async operations
         self._current_flex_data = None
         self._current_force_data = None
-        self.sensor_logger = SensorLog.instance()
+        self.sensor_logger = None
         
         # Initially disable buttons until connection is established
         self.view.set_button_states(False)
@@ -43,7 +42,7 @@ class SensorPresenter:
                 await self.service.stop_force_sensor_notify()
                 
                 # Then stop logging if active
-                if self.sensor_logger.is_logging:
+                if self.sensor_logger and self.sensor_logger.is_logging:
                     self.sensor_logger.stop_logging()
                     # Reset log button state
                     self.view.reset_log_button()
@@ -55,6 +54,10 @@ class SensorPresenter:
                 self._current_force_data = None
         except Exception as e:
             print(f"Error stopping sensor notifications: {e}")
+            
+    def set_logger(self, logger):
+        """Set sensor logger instance"""
+        self.sensor_logger = logger
 
     async def _handle_flex_update(self, sender, flex_data):
         """Handle flex sensor data updates
@@ -70,7 +73,7 @@ class SensorPresenter:
                 self.view.update_flex_sensor(i, value)
                 
             # Write to log if force data is also available
-            if self._current_force_data:
+            if self._current_force_data and self.sensor_logger and self.sensor_logger.is_logging:
                 self.sensor_logger.write_csv(flex_data.values, self._current_force_data.value)
 
     async def _handle_force_update(self, sender, force_data):
@@ -86,6 +89,5 @@ class SensorPresenter:
             self.view.update_force_sensor(force_data.value)
             
             # Write to log if flex data is also available
-            if self._current_flex_data:
+            if self._current_flex_data and self.sensor_logger and self.sensor_logger.is_logging:
                 self.sensor_logger.write_csv(self._current_flex_data.values, force_data.value)
-
